pyproject.toml
[tool.poetry]
name = "scssg"
version = "0.1.0"
description = ""
authors = ["Noah Roberts <noah@noahsroberts.com>"]
license = "GPLv3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
markdown2 = "^2.5.3"
jinja2 = "^3.1.5"
pyyaml = "^6.0.2"

[tool.poetry.scripts]
# build

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
